True A.I. implementation
Added some elements of basic A.I. - more to come in a near future. For now, in full "auto" method the app will progressively implement a map of your preferences and remember them to approximate your preferences under similar circumstances. 
For now the app learns based on outside temp and a basic preset map that gets implemented over time. In the future I'll implement more parameters, such as differentiated humidity values (inside humidity : outside humidity : outside temperature : target amplitude), which will allow the app to be more thorough. 

Note that these features are already available with the app called "A.I. Thermostat manager" and supposed to work in parallel with this one, using the dimmer to share computed results. However, this app takes too much ressources as is and it is to be entirely reworked and implemented into Thermostat Manager directly, as I have started doing so with this new learning algorithm. In order to achieve complete learning abilities, I'll probably use the new file feature in Hubitat to store the big data needed to that effect.